# PhaseIV/specs/001-todo-web-app/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo Web App API
  description: API for managing user tasks in the Todo Web App
  version: 1.0.0
  contact:
    name: Todo Web App Team
servers:
  - url: https://api.todowebapp.com/v1
    description: Production server
  - url: http://localhost:8000/v1
    description: Development server
paths:
...
  /tasks:
    get:
      summary: Get user's tasks
      description: Retrieve all tasks belonging to the authenticated user
      tags:
        - Tasks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successfully retrieved tasks
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/TaskResponse'
        '401':
          description: Unauthorized - invalid or missing JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - user not allowed to access these resources
        '500':
          description: Internal server error

    post:
      summary: Create a new task
      description: Creates a new task for the authenticated user
      tags:
        - Tasks
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
      responses:
        '201':
          description: Task successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or missing JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - user not allowed to create tasks
        '500':
          description: Internal server error
  /tasks/{taskId}:
    get:
      summary: Get a specific task
      description: Retrieve a specific task by ID for the authenticated user
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the task to retrieve
      responses:
        '200':
          description: Successfully retrieved task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          description: Unauthorized - invalid or missing JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - user not allowed to access this task
        '404':
          description: Task not found
        '500':
          description: Internal server error

    put:
      summary: Update a task
      description: Updates an existing task for the authenticated user
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the task to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdateRequest'
      responses:
        '200':
          description: Task successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or missing JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - user not allowed to update this task
        '404':
          description: Task not found
        '500':
          description: Internal server error

    delete:
      summary: Delete a task
      description: Deletes a task for the authenticated user
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the task to delete
      responses:
        '204':
          description: Task successfully deleted
        '401':
          description: Unauthorized - invalid or missing JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - user not allowed to delete this task
        '404':
          description: Task not found
        '500':
          description: Internal server error
  /tasks/{taskId}/complete:
    patch:
      summary: Mark task as complete/incomplete
      description: Toggle the completion status of a task
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the task to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                completed:
                  type: boolean
                  description: Set task completion status
              required:
                - completed
      responses:
        '200':
          description: Task completion status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or missing JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - user not allowed to update this task
        '404':
          description: Task not found
        '500':
          description: Internal server error
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token required for authentication
  schemas:
...
    TaskCreateRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
          description: Task title
        description:
          type: string
          maxLength: 1000
          description: Task description (optional)
        due_date:
          type: string
          format: date-time
          description: Task due date (optional)
      required:
        - title
    TaskUpdateRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
          description: Task title
        description:
          type: string
          maxLength: 1000
          description: Task description
        due_date:
          type: string
          format: date-time
          description: Task due date (optional)
    TaskResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Task's unique identifier
        title:
          type: string
          description: Task title
        description:
          type: string
          description: Task description
        status:
          type: string
          enum: [pending, completed]
          description: Task status
        user_id:
          type: string
          format: uuid
          description: ID of user who owns this task
        created_at:
          type: string
          format: date-time
          description: When the task was created
        updated_at:
          type: string
          format: date-time
          description: When the task was last updated
        due_date:
          type: string
          format: date-time
          description: Task due date (optional)
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details